refactor Map outside of Venue for reuse
